Version 0.6.0
- Removed seed data
- Created exchange rate file templates as backup
- Improved exchange rate processing
- Added API response codes to readme

+++ src/main/java/fi/free/apilayer/ApiLayerService.java
@@ -1,5 +1,6 @@
 package fi.free.apilayer;
 
+import java.io.IOException;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Map;
@@ -8,6 +9,8 @@ import java.util.Set;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.core.io.ClassPathResource;
+import org.springframework.core.io.Resource;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpMethod;
@@ -15,6 +18,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestClientException;
 import org.springframework.web.client.RestTemplate;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
 
 import fi.free.api.controllers.ApiController;
 import fi.free.models.ExchangeRate;
@@ -31,7 +36,8 @@ public class ApiLayerService {
 
 	private final ExchangeRateRepository exchangeRateRepository;
 
-	public ApiLayerService(@Value("${fi.free.apilayer.apiKey}") String apiLayerApiKey, ExchangeRateRepository exchangeRateRepository) {
+	public ApiLayerService(@Value("${fi.free.apilayer.apiKey}") String apiLayerApiKey,
+			ExchangeRateRepository exchangeRateRepository) {
 		this.apilayer_apikey = apiLayerApiKey;
 		this.exchangeRateRepository = exchangeRateRepository;
 	}
@@ -39,29 +45,59 @@ public class ApiLayerService {
 	/**
 	 * Gets latest exchange rates for currencies and stores them to database.
 	 * 
-	 * @param base Currency code used to as base currency
-	 * @param symbols All currency codes that must be included in the exchange rate result
+	 * @param base    Currency code used to as base currency
+	 * @param symbols All currency codes that must be included in the exchange rate
+	 *                result
 	 */
 	public void getLatestExchangeRatesFromApiLayer(String base, String[] symbols) {
 		log.info("Get latest exchange rates for currency {}", base);
+		ExchangeRateResponse response = null;
 		try {
-			Set<ExchangeRate> exchangeRates = new HashSet<ExchangeRate>();
-			ExchangeRateResponse response = fetchLatestExchangeRates(base, symbols);
-			exchangeRateRepository.deleteAllByBase(base);
-			response.getRates().forEach((key, value) -> {
-				ExchangeRate exchangeRate = new ExchangeRate();
-				exchangeRate.setBase(response.getBase());
-				exchangeRate.setDate(response.getDate());
-				exchangeRate.setExchangeRate(key);
-				exchangeRate.setExchangeValue(value);
-				exchangeRate.setTimestamp(response.getTimestamp());
-				exchangeRates.add(exchangeRate);
-			});
-			exchangeRateRepository.saveAll(exchangeRates);
-			exchangeRateRepository.flush();
+			response = fetchLatestExchangeRates(base, symbols);
 		} catch (RestClientException e) {
 			log.warn("Error while fetching new exchange rates. Reason: {}", e.getMessage());
+			// This block is meant to be backup to simulate real SUCCESSFULL response from apilayer...which seems to be constantly exceeding the usage limits.
+			response = loadLocalExchangeRateResponse(base.toLowerCase() + ".json");
+			log.debug("{} {} {}", response.getBase(), response.getDate(), response.getRates().toString());
 		}
+		if(response == null)
+			return;
+		Set<ExchangeRate> exchangeRates = convertExchangeRateResponseToExchangeRates(response);
+		exchangeRateRepository.deleteAllByBase(base);
+		exchangeRateRepository.saveAll(exchangeRates);
+		exchangeRateRepository.flush();
+	}
+
+	private ExchangeRateResponse loadLocalExchangeRateResponse(String fileName) {
+		log.info("Loading backup template file {}", fileName);
+		Resource templateFile = new ClassPathResource(fileName);
+		ObjectMapper mapper = new ObjectMapper();
+		try {
+			return mapper.readValue(templateFile.getInputStream(), ExchangeRateResponse.class);
+		} catch (IOException e) {
+			log.warn("Failed to load local ExchangeRate template file");
+		}
+		return null;
+	}
+
+	/**
+	 * Converts ExchangeRateResponse to ExchangeRate models
+	 * 
+	 * @param response
+	 * @return Set<ExchangeRate>
+	 */
+	private Set<ExchangeRate> convertExchangeRateResponseToExchangeRates(ExchangeRateResponse response) {
+		Set<ExchangeRate> exchangeRates = new HashSet<ExchangeRate>();
+		response.getRates().forEach((key, value) -> {
+			ExchangeRate exchangeRate = new ExchangeRate();
+			exchangeRate.setBase(response.getBase());
+			exchangeRate.setDate(response.getDate());
+			exchangeRate.setExchangeRate(key);
+			exchangeRate.setExchangeValue(value);
+			exchangeRate.setTimestamp(response.getTimestamp());
+			exchangeRates.add(exchangeRate);
+		});
+		return exchangeRates;
 	}
 
 	/**
@@ -71,8 +107,7 @@ public class ApiLayerService {
 	 * @param symbols Exchange rate currencies
 	 * @return ExchangeRateResponse Latest exchange rate data
 	 */
-	private ExchangeRateResponse fetchLatestExchangeRates(String base, String[] symbols)
-			throws RestClientException {
+	private ExchangeRateResponse fetchLatestExchangeRates(String base, String[] symbols) throws RestClientException {
 		log.info("Sending HTTP request to ApiLayer URL {}", apilayer_latest_url);
 		RestTemplate request = new RestTemplate();
 		final HttpHeaders headers = new HttpHeaders();
@@ -82,9 +117,10 @@ public class ApiLayerService {
 		queryParams.put("base", base);
 		queryParams.put("symbols", symbols.toString());
 		ResponseEntity<ExchangeRateResponse> response = null;
-		response = request.exchange(apilayer_latest_url, HttpMethod.GET, entity, ExchangeRateResponse.class, queryParams);
+		response = request.exchange(apilayer_latest_url, HttpMethod.GET, entity, ExchangeRateResponse.class,
+				queryParams);
 		log.info("Response from ApiLayer {} {}", response.getStatusCodeValue(), response.getStatusCode());
 		return response.getBody();
 	}
-	
+
 }
